initial checking for using scm source implementations specifically for on_pull_request get_source_branch
