+++ github/impls.groovy
@@ -1,0 +1,5 @@
+
+def call(){
+    String impl = config.source_type
+    return getBinding().getStep(impl)
+}
